Node color literal in service has been externalised to constant Values is drop-down made to be shown in ascending order DomainValue Id is now long so that multiple masters could be accommodated Right bar title made to describe the scenario clearly Only relevant buttons are shown now in right bar Bug fixed: To be searched node becomes to be saved node after action becomes save in-between on a different node Suppress Invalid person -1 message on double click Zoom on double click disabled

+++ src/main/java/org/sakuram/relation/valueobject/AttributeValueVO.java
@@ -4,7 +4,7 @@ import java.sql.Date;
 
 public class AttributeValueVO {
 	private long id;
-	private int attributeDvId;
+	private long attributeDvId;
 	private String attributeName;
 	private String attributeValue;
 	private boolean isValueAccurate;
@@ -19,11 +19,11 @@ public class AttributeValueVO {
 		this.id = id;
 	}
 	
-	public int getAttributeDvId() {
+	public long getAttributeDvId() {
 		return attributeDvId;
 	}
 
-	public void setAttributeDvId(int attributeDvId) {
+	public void setAttributeDvId(long attributeDvId) {
 		this.attributeDvId = attributeDvId;
 	}
